% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/issues.R
\name{createIssue}
\alias{createIssue}
\title{Creates a new issue}
\usage{
createIssue(repos, title, body, assignee = NULL, milestone = NULL,
  labels = NULL, agrep = 0.1)
}
\arguments{
\item{repos}{A GitHubRepository object.}

\item{title}{Character string.}

\item{body}{Character string.}

\item{assignee}{(optional) Character string.}

\item{milestone}{(optional) Character vector.}

\item{labels}{(optional) Character vector.}

\item{agrep}{A numeric in [0,1] used to test for already existing
issues in order to avoid duplicates.}
}
\value{
Returns a named logical vector.
}
\description{
Creates a new issue
}

