% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/labels.R
\name{createLabel}
\alias{createLabel}
\title{Creates a new label}
\usage{
createLabel(label, color, owner, repo)
}
\arguments{
\item{label}{Character string.}

\item{color}{Character string.}

\item{owner}{Character string.}

\item{repo}{Character string.}
}
\value{
Returns a named logical vector.
}
\description{
Creates a new label
}

