% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/issues.R
\name{hasIssues}
\alias{hasIssues}
\title{Checks existance of issues}
\usage{
hasIssues(repos, titles, agrep = 0)
}
\arguments{
\item{repos}{A GitHubRepository object.}

\item{titles}{Character vector.}

\item{agrep}{A numeric in [0,1] controlling amount of discrepance allowed.}
}
\value{
Returns a named logical vector.
}
\description{
Checks whether a set of issues exists or not by
scanning issue titles.
}

