% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/labels.R
\name{hasLabels}
\alias{hasLabels}
\title{Checks existance of labels}
\usage{
hasLabels(labels, owner, repo)
}
\arguments{
\item{labels}{Character vector.}

\item{owner}{Character string.}

\item{repo}{Character string.}
}
\value{
Returns a named logical vector.
}
\description{
Checks whether a set of labels exists or not.
}

