% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/issues.R
\name{listIssues}
\alias{listIssues}
\title{List issues}
\usage{
listIssues(owner, repo)
}
\arguments{
\item{owner}{Character string.}

\item{repo}{Character string.}
}
\value{
Returns a character vector.
}
\description{
Retrievs all issues.
}

