% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/labels.R
\name{listLabels}
\alias{listLabels}
\title{List labels}
\usage{
listLabels(repos)
}
\arguments{
\item{repos}{A GitHubRepository object.}
}
\value{
Returns a character vector.
}
\description{
Retrievs all labels.
}

