% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/repos.R
\name{repos}
\alias{repos}
\title{Creates a reference to a GitHub repository}
\usage{
repos(owner, repo, auth = NULL)
}
\arguments{
\item{owner}{Character string.}

\item{repo}{Character string.}

\item{auth}{(optional) Character string.}
}
\value{
Returns a named list.
}
\description{
Creates a reference to a GitHub repository
}

