% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/issues.R
\name{hasIssues}
\alias{hasIssues}
\title{Checks existance of issues}
\usage{
hasIssues(titles, agrep = 0, owner, repo)
}
\arguments{
\item{titles}{Character vector.}

\item{agrep}{A numeric in [0,1] controlling amount of discrepance allowed.}

\item{owner}{Character string.}

\item{repo}{Character string.}
}
\value{
Returns a named logical vector.
}
\description{
Checks whether a set of issues exists or not by
scanning issue titles.
}

