% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/issues.R
\name{listIssues}
\alias{listIssues}
\title{List issues}
\usage{
listIssues(repos)
}
\arguments{
\item{repos}{A GitHubRepository object.}
}
\value{
Returns a character vector.
}
\description{
Retrievs all issues.
}

