% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/labels.R
\name{listLabels}
\alias{listLabels}
\title{List labels}
\usage{
listLabels(owner, repo)
}
\arguments{
\item{owner}{Character string.}

\item{repo}{Character string.}
}
\value{
Returns a character vector.
}
\description{
Retrievs all labels.
}

